Moving classes to their new places.
