+++ graylog-storage-elasticsearch6/src/main/java/org/graylog/storage/elasticsearch6/views/searchtypes/pivot/series/ESMinHandler.java
@@ -14,20 +14,18 @@
  * You should have received a copy of the GNU General Public License
  * along with Graylog.  If not, see <http://www.gnu.org/licenses/>.
  */
-package org.graylog.plugins.views.search.elasticsearch.searchtypes.pivot.series;
+package org.graylog.storage.elasticsearch6.views.searchtypes.pivot.series;
 
 import io.searchbox.core.SearchResult;
 import io.searchbox.core.search.aggregation.MinAggregation;
 import org.elasticsearch.search.aggregations.AggregationBuilder;
 import org.elasticsearch.search.aggregations.AggregationBuilders;
 import org.elasticsearch.search.aggregations.metrics.min.MinAggregationBuilder;
-import org.graylog.plugins.views.search.elasticsearch.ESGeneratedQueryContext;
-import org.graylog.plugins.views.search.elasticsearch.searchtypes.pivot.ESPivot;
-import org.graylog.plugins.views.search.elasticsearch.searchtypes.pivot.ESPivotSeriesSpecHandler;
+import org.graylog.storage.elasticsearch6.views.ESGeneratedQueryContext;
+import org.graylog.storage.elasticsearch6.views.searchtypes.pivot.ESPivot;
+import org.graylog.storage.elasticsearch6.views.searchtypes.pivot.ESPivotSeriesSpecHandler;
 import org.graylog.plugins.views.search.searchtypes.pivot.Pivot;
 import org.graylog.plugins.views.search.searchtypes.pivot.series.Min;
-import org.graylog.plugins.views.search.elasticsearch.searchtypes.pivot.ESPivot;
-import org.graylog.plugins.views.search.elasticsearch.searchtypes.pivot.ESPivotSeriesSpecHandler;
 
 import javax.annotation.Nonnull;
 import java.util.Optional;
