+++ graylog-storage-elasticsearch6/src/main/java/org/graylog/storage/elasticsearch6/views/searchtypes/pivot/series/ESStdDevHandler.java
@@ -14,40 +14,38 @@
  * You should have received a copy of the GNU General Public License
  * along with Graylog.  If not, see <http://www.gnu.org/licenses/>.
  */
-package org.graylog.plugins.views.search.elasticsearch.searchtypes.pivot.series;
+package org.graylog.storage.elasticsearch6.views.searchtypes.pivot.series;
 
 import io.searchbox.core.SearchResult;
 import io.searchbox.core.search.aggregation.ExtendedStatsAggregation;
 import org.elasticsearch.search.aggregations.AggregationBuilder;
 import org.elasticsearch.search.aggregations.AggregationBuilders;
 import org.elasticsearch.search.aggregations.metrics.stats.extended.ExtendedStatsAggregationBuilder;
-import org.graylog.plugins.views.search.elasticsearch.ESGeneratedQueryContext;
-import org.graylog.plugins.views.search.elasticsearch.searchtypes.pivot.ESPivot;
-import org.graylog.plugins.views.search.elasticsearch.searchtypes.pivot.ESPivotSeriesSpecHandler;
+import org.graylog.storage.elasticsearch6.views.ESGeneratedQueryContext;
+import org.graylog.storage.elasticsearch6.views.searchtypes.pivot.ESPivot;
+import org.graylog.storage.elasticsearch6.views.searchtypes.pivot.ESPivotSeriesSpecHandler;
 import org.graylog.plugins.views.search.searchtypes.pivot.Pivot;
-import org.graylog.plugins.views.search.searchtypes.pivot.series.Variance;
-import org.graylog.plugins.views.search.elasticsearch.searchtypes.pivot.ESPivotSeriesSpecHandler;
+import org.graylog.plugins.views.search.searchtypes.pivot.series.StdDev;
 
 import javax.annotation.Nonnull;
 import java.util.Optional;
 import java.util.stream.Stream;
 
-public class ESVarianceHandler extends ESPivotSeriesSpecHandler<Variance, ExtendedStatsAggregation> {
+public class ESStdDevHandler extends ESPivotSeriesSpecHandler<StdDev, ExtendedStatsAggregation> {
     @Nonnull
     @Override
-    public Optional<AggregationBuilder> doCreateAggregation(String name, Pivot pivot, Variance varianceSpec, ESPivot searchTypeHandler, ESGeneratedQueryContext queryContext) {
-        final ExtendedStatsAggregationBuilder variance = AggregationBuilders.extendedStats(name).field(varianceSpec.field());
-        record(queryContext, pivot, varianceSpec, name, ExtendedStatsAggregation.class);
-        return Optional.of(variance);
+    public Optional<AggregationBuilder> doCreateAggregation(String name, Pivot pivot, StdDev stddevSpec, ESPivot searchTypeHandler, ESGeneratedQueryContext queryContext) {
+        final ExtendedStatsAggregationBuilder stddev = AggregationBuilders.extendedStats(name).field(stddevSpec.field());
+        record(queryContext, pivot, stddevSpec, name, ExtendedStatsAggregation.class);
+        return Optional.of(stddev);
     }
 
     @Override
-    public Stream<Value> doHandleResult(Pivot pivot, Variance pivotSpec,
+    public Stream<Value> doHandleResult(Pivot pivot, StdDev pivotSpec,
                                         SearchResult searchResult,
-                                        ExtendedStatsAggregation varianceAggregation,
+                                        ExtendedStatsAggregation stddevAggregation,
                                         ESPivot searchTypeHandler,
                                         ESGeneratedQueryContext esGeneratedQueryContext) {
-        return Stream.of(ESPivotSeriesSpecHandler.Value.create(pivotSpec.id(), Variance.NAME, varianceAggregation.getVariance()));
+        return Stream.of(ESPivotSeriesSpecHandler.Value.create(pivotSpec.id(), StdDev.NAME, stddevAggregation.getStdDeviation()));
     }
 }
-
