+++ graylog2-server/src/test/java/org/graylog/plugins/views/search/export/LegacyChunkDecoratorTest.java
@@ -18,7 +18,7 @@ package org.graylog.plugins.views.search.export;
 
 import org.assertj.core.groups.Tuple;
 import org.graylog.plugins.views.search.elasticsearch.ElasticsearchQueryString;
-import org.graylog.plugins.views.search.elasticsearch.searchtypes.pivot.LegacyDecoratorProcessor;
+import org.graylog.storage.elasticsearch6.views.searchtypes.pivot.LegacyDecoratorProcessor;
 import org.graylog2.plugin.indexer.searches.timeranges.AbsoluteRange;
 import org.graylog2.rest.models.messages.responses.ResultMessageSummary;
 import org.graylog2.rest.resources.search.responses.SearchResponse;
